vim-patch:8.1.0857: indent functionality is not separated
Problem:    Ignore functionality is not separated.
Solution:   Move indent functionality into a new file. (Yegappan Lakshmanan,
            closes vim/vim#3886)

https://github.com/vim/vim/commit/4b47162ccede0b6d9cbb9473ad870220a24fbf54

----

Partial port of v8.1.2127 by porting directly to indent_c.c,
not indent.c.

----
